back/requirements.txt
absl-py==2.1.0
asttokens==2.4.1
blinker==1.7.0
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
contourpy==1.2.0
cycler==0.12.1
decorator==5.1.1
exceptiongroup==1.2.0
executing==2.0.1
filelock==3.13.3
Flask==3.0.2
Flask-Cors==4.0.0
fonttools==4.50.0
fsspec==2024.3.1
grpcio==1.62.1
idna==3.6
ipython==8.23.0
itsdangerous==2.1.2
jedi==0.19.1
Jinja2==3.1.3
kiwisolver==1.4.5
Markdown==3.6
MarkupSafe==2.1.5
matplotlib==3.8.3
matplotlib-inline==0.1.6
mpmath==1.3.0
networkx==3.2.1
numpy==1.26.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.19.3
nvidia-nvjitlink-cu12==12.4.99
nvidia-nvtx-cu12==12.1.105
opencv-python==4.9.0.80
opencv-python-headless==4.9.0.80
packaging==24.0
pandas==2.2.1
parso==0.8.3
pexpect==4.9.0
pillow==10.3.0
prompt-toolkit==3.0.43
protobuf==5.26.1
psutil==5.9.8
ptyprocess==0.7.0
pure-eval==0.2.2
Pygments==2.17.2
pyparsing==3.1.2
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.1
requests==2.31.0
scipy==1.12.0
seaborn==0.13.2
sentry-sdk==1.44.0
setuptools==59.6.0
six==1.16.0
stack-data==0.6.3
supervision==0.15.0
sympy==1.12
tensorboard==2.16.2
tensorboard-data-server==0.7.2
thop==0.1.1.post2209072238
torch==2.2.2
torchvision==0.17.2
tqdm==4.66.2
traitlets==5.14.2
triton==2.2.0
typing_extensions==4.10.0
tzdata==2024.1
ultralytics==8.0.35
urllib3==2.2.1
wcwidth==0.2.13
Werkzeug==3.0.2
wheel==0.43.0
